LayoutTests/fast/css/round-trip-values-expected.txt

For each input on the left, this table lists what happens when you pass it into a CSS property and retrieve its value back. The third column indicates whether the value was round-tripped faithfully.

Basic floats
'0.0001'	0.0001	pass
0.0001	0.0001	pass
'123456.123456'	123456.123456	pass
'1234567.1234567'	1234567.123457	pass
'12345678.12345678'	12345678.123457	pass
Trailing zeros
'0.00100000'	0.001	pass
'0.001000001'	0.001	pass
'0.12345000001'	0.12345	pass
'0.12304567'	0.123046	pass
'0.12340567'	0.123406	pass
'0.12345067'	0.123451	pass
'0.12345607'	0.123456	pass
'0.12345670'	0.123457	pass
Repeating decimals
1/3	0.333333	pass
123 + 1/3	123.333333	pass
13/99	0.131313	pass
123 + 13/99	123.131313	pass
100/999	0.1001	pass
123 + 100/999	123.1001	pass
Large numbers
12345678	12345678	pass
123456789	123456789	pass
1234567890	1234567890	pass
12345678901	12345678901	pass
123456789012	123456789012	pass
1234567890123	1234567890123	pass
12345678901234	12345678901234	pass
123456789012345	123456789012345	pass
1234567890123456	1234567890123456	pass
12345678901234567	12345678901234568	pass
Weird numbers
Number.NaN	12345678901234568	pass
1/0	12345678901234568	pass
Math.sqrt(-1)	12345678901234568	pass
1/0.9999	1.0001	pass
1/0.99999	1.00001	pass
1/0.999999	1.000001	pass
1/0.9999999	1	pass
1/0.99999999	1	pass
